# target/classes/api.yml
openapi: 3.0.3
info:
  title: Credit Service API
  version: 4.0.0
  description: >
    Credit products API.
    - Loans (PERSONAL/BUSINESS) live under /credits
    - Credit card lines live under /credit-lines
servers:
  - url: http://localhost:8585
...
paths:
...
  /credit-lines:
    post:
      tags: [ CreditLines ]
      summary: Create credit line
      description: Creates a credit line used by credit cards. # Crea línea para tarjeta
      operationId: createCreditLine
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreditLineRequest' }
      responses:
        '201':
          description: Credit line created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreditLineResponse' }

    get:
      tags: [ CreditLines ]
      summary: List credit lines
      operationId: listCreditLines
      parameters:
        - in: query
          name: customerId
          schema: { type: string }
        - in: query
          name: status
          schema: { $ref: '#/components/schemas/CreditStatus' }
      responses:
        '200':
          description: Credit line list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CreditLineResponse' }
  /credit-lines/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [ CreditLines ]
      summary: Get credit line by ID
      operationId: getCreditLine
      responses:
        '200':
          description: Credit line
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreditLineResponse' }
        '404': { $ref: '#/components/responses/NotFound' }

    patch:
      tags: [ CreditLines ]
      summary: Partial update (credit line)
      operationId: patchCreditLine
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema: { $ref: '#/components/schemas/CreditLineUpdate' }
      responses:
        '200':
          description: Credit line updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreditLineResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }

    put:
      tags: [ CreditLines ]
      summary: Replace credit line
      operationId: updateCreditLine
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreditLineRequest' }
      responses:
        '200':
          description: Credit line replaced
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreditLineResponse' }
    delete:
      tags: [ CreditLines ]
      summary: Delete credit line
      operationId: deleteCreditLine
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /credit-lines/{id}/limit:
    patch:
      tags: [ CreditLines ]
      summary: Adjust credit line limit
      description: New limit must be ≥ current balance. # Límite ≥ saldo
      operationId: adjustCreditLineLimit
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ newLimit ]
              properties:
                newLimit: { type: number, minimum: 0 }
                reason: { type: string, maxLength: 200 }
      responses:
        '200':
          description: Limit updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreditLineResponse' }
  /credit-lines/{id}/charges:
    post:
      tags: [ CreditLines ]
      summary: Post charge (card purchase)
      description: Validates available amount before posting. # Valida disponible
      operationId: postCreditLineCharge
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreditChargeRequest' }
      responses:
        '201':
          description: Charge posted
          headers:
            Location:
              description: Movement URI
              schema: { type: string, format: uri }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreditMovement' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /credit-lines/{id}/payments:
    post:
      tags: [ CreditLines ]
      summary: Post payment to the credit line
      operationId: postCreditLinePayment
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreditPaymentRequest' }
      responses:
        '201':
          description: Payment posted
          headers:
            Location:
              description: Movement URI
              schema: { type: string, format: uri }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreditMovement' }

#  /credit-lines/{id}/balance:
#    get:
#      tags: [ CreditLines ]
#      summary: Get credit line balance
#      description: Returns limit, balance and available. # Saldos de línea
#      operationId: getCreditLineBalance
#      parameters:
#        - in: path
#          name: id
#          required: true
#          schema: { type: string }
#      responses:
#        '200':
#          description: OK
#          content:
#            application/json:
#              schema: { $ref: '#/components/schemas/CreditBalance' }
#        '404': { $ref: '#/components/responses/NotFound' }
#
#  /credit-lines/{id}/movements:
#    get:
#      tags: [ CreditLines ]
#      summary: List credit line movements
#      description: Returns movements for the credit line. # Movimientos de tarjeta
#      operationId: listCreditLineMovements
#      parameters:
#        - in: path
#          name: id
#          required: true
#          schema: { type: string }
#        - in: query
#          name: type
#          schema:
#            type: string
#            enum: [ CHARGE, PAYMENT, INTEREST, FEE, REVERSAL ]
#        - in: query
#          name: limit
#          schema: { type: integer, minimum: 1, maximum: 100, default: 10 } # Últimos 10 por defecto
#      responses:
#        '200':
#          description: Movement list
#          content:
#            application/json:
#              schema:
#                type: array
#                items: { $ref: '#/components/schemas/CreditMovement' }
  /credit-lines/{id}/close:
    post:
      tags: [ CreditLines ]
      summary: Close credit line (logical delete)
      description: Requires balance=0. # Requiere saldo 0
      operationId: closeCreditLine
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Credit line closed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreditLineResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }

  #Common / Cross Product
...
components:
  responses:
    BadRequest:
      description: Validation or bad request.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Conflict with business rules or resource state.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    UnprocessableEntity:
      description: Semantically valid but violates domain constraints
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    CreditStatus:
      type: string
      description: Estado del crédito.
      enum: [ ACTIVE, CLOSED, DELINQUENT ]
...
    CreditBalance:
      type: object
      description: Product balances (limit, used, available). # Saldos
      required: [ limit, balance, available ]
      additionalProperties: false
      properties:
        limit: { type: number, minimum: 0 }
        balance: { type: number, minimum: 0 }
        available: { type: number }
    CreditMovement:
      type: object
      description: Movement (charge, payment, interest, fee or reversal). # Movimiento
      required: [ id, creditId, type, amount, txnAt ]
      additionalProperties: false
      properties:
        id: { type: string, description: Movement id }
        creditId: { type: string, description: Product id (loan or credit line) } # Id del producto
        type:
          type: string
          enum: [ CHARGE, PAYMENT, INTEREST, FEE, REVERSAL ]
        amount: { type: number }
        txnAt: { type: string, format: date-time }
        postedAt: { type: string, format: date-time }
        runningBalance: { type: number }
        channel: { type: string }
    CreditPaymentRequest:
      type: object
      description: Payment data (owner or third-party). # Datos de pago
      required: [ amount ]
      additionalProperties: false
      properties:
        payerCustomerId: { type: string, description: Payer customer id (optional). }
        amount: { type: number, minimum: 0.01 }
        channel:
          type: string
          enum: [ CASH, TRANSFER, CARD ]
        note: { type: string, maxLength: 200 }
    CreditChargeRequest:
      type: object
      description: Charge data (card purchase). # Datos de cargo (compra)
      required: [ amount ]
      additionalProperties: false
      properties:
        amount: { type: number, minimum: 0.01 }
        merchant: { type: string, maxLength: 120 }
        channel: { type: string, default: CARD }
        note: { type: string, maxLength: 200 }
...
    Error:
      type: object
      description: RFC 7807 Problem Details. # Estructura de error
      additionalProperties: false
      properties:
        type: { type: string, example: 'https://errors.bank.com/business-rule' }
        title: { type: string, example: 'Business rule violation' }
        status: { type: integer, example: 409 }
        detail: { type: string, example: 'Customer already has an active PERSONAL loan.' }
        instance: { type: string, example: '/credits/12345' }
        code: { $ref: '#/components/schemas/ErrorCode' }
        errors:
          type: array
          items: { $ref: '#/components/schemas/FieldError' }
    FieldError:
      type: object
      description: Field validation error. # Error por campo
      additionalProperties: false
      properties:
        field: { type: string, example: 'interestAnnual' }
        message: { type: string, example: 'Must be greater or equal than 0' }
        code: { type: string, example: 'Min' }
    ErrorCode:
      type: string
      description: Business code. # Código interno
      enum:
        - CREDIT_PERSONAL_LIMIT
        - CREDIT_DELINQUENT_LOCK
        - CREDIT_CARD_LIMIT_EXCEEDED
        - CREDIT_INVALID_STATE
        - CREDIT_NOT_FOUND
        - CUSTOMER_NOT_FOUND
        - VALIDATION_ERROR
        - SERVICE_UNAVAILABLE

  # Credit Lines
    CreditLineResponse:
      type: object
      description: Credit line stored in the system. # Línea de crédito
      required: [id, customerId, limit, balance, interestAnnual, status, createdAt]
      properties:
        id: { type: string, readOnly: true }
        customerId: { type: string, readOnly: true }
        limit: { type: number, minimum: 0 }
        balance: { type: number, minimum: 0, readOnly: true }
        interestAnnual: { type: number, minimum: 0 }
        billingCycleDay:
          type: integer
          minimum: 1
          maximum: 28
          description: Statement closing day.
        paymentDueDay:
          type: integer
          minimum: 1
          maximum: 28
          description: Payment due day.
        status: { $ref: '#/components/schemas/CreditStatus', readOnly: true }
        currency: { type: string, example: PEN }
        closedAt: { type: string, format: date-time, readOnly: true }
        closeReason: { type: string, readOnly: true }
        createdAt: { type: string, format: date-time, readOnly: true }
        updatedAt: { type: string, format: date-time, readOnly: true }
    CreditLineRequest:
      type: object
      description: Credit line creation payload. # Payload de creación de línea
      additionalProperties: false
      properties:
        customerId: { type: string }
        limit: { type: number, minimum: 0 }
        interestAnnual: { type: number, minimum: 0 }
        billingCycleDay: { type: integer, minimum: 1, maximum: 28 }
        paymentDueDay: { type: integer, minimum: 1, maximum: 28 }
        currency: { type: string, example: PEN, default: PEN }
      required: [customerId, limit, interestAnnual, currency, billingCycleDay, paymentDueDay]
    CreditLineUpdate:
      type: object
      description: Allowed fields for partial update. # Campos permitidos en PATCH
      additionalProperties: false
      properties:
        interestAnnual: { type: number, minimum: 0 }
        billingCycleDay: { type: integer, minimum: 1, maximum: 28 }
        paymentDueDay: { type: integer, minimum: 1, maximum: 28 }
...